Delete item action added+minor fix

+++ src/net/epam/study/controller/commands/impl/AdminOrderDelete.java
@@ -2,6 +2,7 @@ package net.epam.study.controller.commands.impl;
 
 import net.epam.study.controller.commands.Command;
 import net.epam.study.dao.OrderDelete;
+import net.epam.study.service.CheckSession;
 
 import javax.servlet.RequestDispatcher;
 import javax.servlet.ServletException;
@@ -12,12 +13,16 @@ import java.io.IOException;
 public class AdminOrderDelete implements Command {
     @Override
     public void execute(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
-        String id = request.getParameter("id");
-        String sql = "delete from orders where id ='" + id + "'";
-        if (id != null){
-            OrderDelete.delete(sql);
+        if (!CheckSession.checkSession(request, response)) {
+            response.sendRedirect("Controller?command=gotologinpage");
+        } else {
+            String id = request.getParameter("id");
+            String sql = "delete from orders where id ='" + id + "'";
+            if (id != null) {
+                OrderDelete.delete(sql);
+            }
+            RequestDispatcher requestDispatcher = request.getRequestDispatcher("/WEB-INF/jsp/admin-indexPage.jsp");
+            requestDispatcher.forward(request, response);
         }
-        RequestDispatcher requestDispatcher = request.getRequestDispatcher("/WEB-INF/jsp/admin-indexPage.jsp");
-        requestDispatcher.forward(request, response);
     }
 }
